Factor out channel creation and reuse channels between the dynamic grpc client and the server reflection client.
